Add fallback for SLE 11 guest machines

+++ plugins/guests/suse/cap/halt.rb
@@ -4,7 +4,11 @@ module VagrantPlugins
       class Halt
         def self.halt(machine)
           begin
-            machine.communicate.sudo("/usr/bin/systemctl poweroff &")
+            if machine.communicate.test("test -e /usr/bin/systemctl")
+              machine.communicate.sudo("/usr/bin/systemctl poweroff &")
+            else
+              machine.communicate.sudo("/sbin/shutdown -h now &")
+            end
           rescue IOError, Vagrant::Errors::SSHDisconnected
             # Do nothing, because it probably means the machine shut down
             # and SSH connection was lost.
